added composed repository, library service interface
