+++ src/main/java/com/ntselishchev/libraryapp/dao/BookDao.java
@@ -4,7 +4,7 @@ import com.ntselishchev.libraryapp.domain.Book;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import reactor.core.publisher.Flux;
 
-public interface BookDao extends ReactiveMongoRepository<Book, String> {
+public interface BookDao extends ReactiveMongoRepository<Book, String>, BookRepository {
 
     Flux<Book> findAll();
