+++ src/main/java/com/ntselishchev/libraryapp/service/LibraryService.java
@@ -1,67 +1,25 @@
 package com.ntselishchev.libraryapp.service;
 
-import com.ntselishchev.libraryapp.dao.AuthorDao;
-import com.ntselishchev.libraryapp.dao.BookDao;
-import com.ntselishchev.libraryapp.dao.BookRepository;
-import com.ntselishchev.libraryapp.dao.GenreDao;
 import com.ntselishchev.libraryapp.domain.Author;
 import com.ntselishchev.libraryapp.domain.Book;
 import com.ntselishchev.libraryapp.domain.Genre;
 import com.ntselishchev.libraryapp.dto.BookDTO;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-@Service
-@RequiredArgsConstructor
-public class LibraryService {
+public interface LibraryService {
 
-    private final BookDao bookDao;
-    private final BookRepository bookRepository;
-    private final AuthorDao authorDao;
-    private final GenreDao genreDao;
+    Mono<Void> addBook(BookDTO bookDto);
 
-    public Mono<Void> addBook(BookDTO bookDto) {
-        return authorDao.findById(bookDto.getAuthorId()).flatMap(a ->
-                genreDao.findById(bookDto.getGenreId()).flatMap(g -> {
-                    Book newBook = new Book(bookDto.getTitle(), a, g);
-                    return bookDao.save(newBook);
-                })
-        ).then();
-    }
+    Mono<Void> deleteBook(String id);
 
-    public Mono<Void> deleteBook(String id) {
-        return bookDao.findById(id)
-                .flatMap(b -> bookRepository.deleteById(id));
-    }
+    Mono<Void> updateBook(BookDTO bookDto);
 
-    public Mono<Void> updateBook(BookDTO bookDto) {
-        return bookDao.findById(bookDto.getId()).flatMap(b ->
-                authorDao.findById(bookDto.getAuthorId()).flatMap(a ->
-                        genreDao.findById(bookDto.getGenreId()).flatMap(g -> {
-                            b.setAuthor(a);
-                            b.setGenre(g);
-                            b.setTitle(bookDto.getTitle());
-                            return bookDao.save(b);
-                        })
-                )
-        ).then();
-    }
+    Flux<Author> getAuthors();
 
-    public Flux<Author> getAuthors() {
-        return authorDao.findAll();
-    }
+    Flux<Genre> getGenres();
 
-    public Flux<Genre> getGenres() {
-        return genreDao.findAll();
-    }
+    Flux<Book> getBooks();
 
-    public Flux<Book> getBooks() {
-        return bookDao.findAll();
-    }
-
-    public Mono<Book> getBook(String id) {
-        return bookDao.findById(id);
-    }
+    Mono<Book> getBook(String id);
 }
